requirements.txt
numpy
torch==2.1.0 --index-url https://download.pytorch.org/whl/cu118
torchvision
tqdm
timm
pycocotools
pytorch_lightning>=2

ipykernel
black
isort
